backend/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via
    #   httpx
    #   openai
    #   starlette
cachetools==5.5.0
    # via google-auth
certifi==2024.12.14
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via uvicorn
colorama==0.4.6
    # via
    #   click
    #   tqdm
datetime==5.5
    # via email-essence (pyproject.toml)
distro==1.9.0
    # via openai
dnspython==2.7.0
    # via
    #   email-validator
    #   pymongo
email-validator==2.2.0
    # via pydantic
fastapi==0.115.6
    # via email-essence (pyproject.toml)
google-ai-generativelanguage==0.6.15
    # via google-generativeai
google-api-core==2.24.0
    # via
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-generativeai
google-api-python-client==2.157.0
    # via
    #   email-essence (pyproject.toml)
    #   google-generativeai
google-auth==2.37.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
    #   google-generativeai
google-auth-httplib2==0.2.0
    # via google-api-python-client
google-auth-oauthlib==1.2.1
    # via email-essence (pyproject.toml)
google-generativeai==0.8.4
    # via email-essence (pyproject.toml)
googleapis-common-protos==1.66.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.70.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.70.0
    # via google-api-core
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.7
    # via httpx
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httpx==0.28.1
    # via openai
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
imapclient==3.0.1
    # via email-essence (pyproject.toml)
jiter==0.8.2
    # via openai
motor==3.7.0
    # via email-essence (pyproject.toml)
oauthlib==3.2.2
    # via requests-oauthlib
openai==1.59.4
    # via email-essence (pyproject.toml)
proto-plus==1.25.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==5.29.2
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-generativeai
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pydantic==2.10.4
    # via
    #   email-essence (pyproject.toml)
    #   fastapi
    #   google-generativeai
    #   openai
    #   pydantic-settings
pydantic-core==2.27.2
    # via pydantic
pydantic-settings==2.7.1
    # via email-essence (pyproject.toml)
pymongo==4.11.1
    # via motor
pyparsing==3.2.1
    # via httplib2
python-dotenv==1.0.1
    # via
    #   email-essence (pyproject.toml)
    #   pydantic-settings
pytz==2025.1
    # via datetime
requests==2.32.3
    # via
    #   google-api-core
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rsa==4.9
    # via google-auth
setuptools==75.8.0
    # via zope-interface
sniffio==1.3.1
    # via
    #   anyio
    #   openai
starlette==0.41.3
    # via fastapi
tenacity==9.0.0
    # via email-essence (pyproject.toml)
tqdm==4.67.1
    # via
    #   google-generativeai
    #   openai
typing==3.10.0.0
    # via email-essence (pyproject.toml)
typing-extensions==4.12.2
    # via
    #   anyio
    #   fastapi
    #   google-generativeai
    #   openai
    #   pydantic
    #   pydantic-core
uritemplate==4.1.1
    # via google-api-python-client
urllib3==2.3.0
    # via requests
uvicorn==0.34.0
    # via email-essence (pyproject.toml)
zope-interface==7.2
    # via datetime
